requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.66.4
flask==3.0.0
scikit-learn
